requirements.txt
addict
yapf
timm
numpy
opencv-python
supervision==0.4.0
paddlepaddle-gpu==2.4.2
paddlenlp==2.5.2